New feature: unsortable lists
Squashed commits:

Made tests for the functionality of disabling sorting within a list. The functionality is not yet implemented so the tests will fail.

Added the 'disableSorting' parameter to the index hbs class. Removed the original test, and made 3 tests for sorting within a disabled sorting list, sorting from a disabled sorting list into a non-disabled sorting one, and sorting from a non-disabled sortinglist into a disabled sorting list.

Made the parameter 'disableSorting'. If this parameter is set to true, a check will be entered during the 'dragOver' event in the drag-sort-item component class. This will make it so the list is not sortable within itself, but still sortable in other lists.

Made a test to drag an item from an unsortable list to a sortable list and back to the unsortable list. The item should not be sorted. This test fails for now. Also fixed some styling issues, and replaced the RSVP promises in addon/utils/trigger.js with await wait().

Removed the disableSorting parameter. Instead made a determineForeignPositionAction action parameter. Changed the tests to reflect this new parameter. The parameter should determine the position of an item when it gets dragged into the unsortable list.

New feature: unsortable lists