Feature: implement drag start action (#45)
* Chore: pin nise

Something that sinon/nise did has caused ember-auto-import to throw an error when trying to import it, so pin nise (which we don't use anyway) to an old version that doesn't blow up.

* Feature: implement drag start action

Implement an onDragStart action that gives consumers access to the event object so they can do things like customize the drag image.

* Drag start action: minor fixes

Co-authored-by: Andrey Mikhaylov (lolmaus) <lolmaus@gmail.com>

Conflicts:
	addon/templates/components/drag-sort-list.hbs
	yarn.lock